Refactor API, improve CLI, add the whole example (#2)
This big PR is here to merge all the API refactoring work, but also and not exhaustively the example and diagrams into `main`.
There remains work to do on the implementation of the Alg4 service, but it will be directly performed over the `main` branch with no reviews.

Tests are a TODO, but not my focus for now.

Co-authored-by: Lucas Tesson <lucastesson@protonmail.com>
Reviewed-on: https://git.cvewatcher.la-ruche.fr:3000/CVEWatcher/godepgraph/pulls/2